refactor routes. Removed unused imports
